requirements.txt
flask==2.3.3
pandas>=2.0.0
plotly==5.17.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
numpy>=1.24.0
